// frontend/src/app/pages/operations/forms/computer/computer-form/computer-form.type.ts
export interface IComputerForm {
  computer_type: string;
  asset_tag: string;
  computer_name: string;
  model_name: string;
  serial_number: string;
  department: string;
  assigned_to: string;
  operating_system: string;
  created_by: number;
  created_date: string;
  active: number;
  include_in_inspection_report: number;
}